fix: add input verifications to the procedures  - refactor declarations and selects  - add new db erros

+++ service/bulk/schema/750$bulk.batch.fetch.sql
@@ -1,6 +1,6 @@
 CREATE OR REPLACE FUNCTION bulk."batch.fetch" (
     "@name" VARCHAR,
-    "@batchStatusId" SMALLINT,
+    "@batchStatusId" INTEGER,
     "@fromDate" TIMESTAMP,
     "@toDate" TIMESTAMP
 )
